Update player screen logic
